+++ src/main/kotlin/io/dotanuki/arw/core/android/AnalysedArtifact.kt
@@ -1,4 +1,4 @@
-package io.dotanuki.arw.core.domain.models
+package io.dotanuki.arw.core.android
 
 data class AnalysedArtifact(
     val applicationId: String,
