+++ src/main/kotlin/io/dotanuki/arw/core/android/AndroidArtifactType.kt
@@ -1,4 +1,4 @@
-package io.dotanuki.arw.core.infrastructure.android
+package io.dotanuki.arw.core.android
 
 enum class AndroidArtifactType {
     AAB,
