+++ src/main/kotlin/io/dotanuki/arw/core/android/AndroidComponent.kt
@@ -1,4 +1,4 @@
-package io.dotanuki.arw.core.domain.models
+package io.dotanuki.arw.core.android
 
 data class AndroidComponent(
     val name: String,
