+++ src/main/kotlin/io/dotanuki/arw/core/android/AndroidComponentType.kt
@@ -1,4 +1,4 @@
-package io.dotanuki.arw.core.domain.models
+package io.dotanuki.arw.core.android
 
 enum class AndroidComponentType {
     APPLICATION,
