+++ src/main/kotlin/io/dotanuki/arw/core/android/AndroidSDKBridge.kt
@@ -1,8 +1,8 @@
-package io.dotanuki.arw.core.infrastructure.android
+package io.dotanuki.arw.core.android
 
 import arrow.core.raise.ensure
-import io.dotanuki.arw.core.domain.errors.ArwError
-import io.dotanuki.arw.core.domain.errors.ErrorAware
+import io.dotanuki.arw.core.errors.ArwError
+import io.dotanuki.arw.core.errors.ErrorAware
 
 object AndroidSDKBridge {
