+++ src/main/kotlin/io/dotanuki/arw/features/baseline/GenerateCommand.kt
@@ -5,11 +5,11 @@ import com.github.ajalt.clikt.core.CliktCommand
 import com.github.ajalt.clikt.parameters.options.flag
 import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.parameters.options.required
-import io.dotanuki.arw.core.domain.errors.ErrorAware
-import io.dotanuki.arw.core.infrastructure.android.AndroidArtifactAnalyser
-import io.dotanuki.arw.core.infrastructure.cli.ErrorReporter
-import io.dotanuki.arw.core.infrastructure.filesystem.ValidatedFile
-import io.dotanuki.arw.features.common.ArtifactBaseline
+import io.dotanuki.arw.core.android.AndroidArtifactAnalyser
+import io.dotanuki.arw.core.cli.ErrorReporter
+import io.dotanuki.arw.core.errors.ErrorAware
+import io.dotanuki.arw.core.filesystem.ValidatedFile
+import io.dotanuki.arw.core.toml.SerializableBaseline
 
 context (BaselineContext)
 class GenerateCommand : CliktCommand(
@@ -30,7 +30,7 @@ class GenerateCommand : CliktCommand(
         val analysed = AndroidArtifactAnalyser.analyse(ValidatedFile(target))
         val outputFile = "${analysed.applicationId}.toml"
 
-        val baseline = ArtifactBaseline.from(analysed)
+        val baseline = SerializableBaseline.from(analysed)
         BaselineWriter.write(baseline, outputFile)
     }
 }
