+++ src/main/kotlin/io/dotanuki/arw/features/comparison/CompareCommand.kt
@@ -5,12 +5,11 @@ import com.github.ajalt.clikt.core.CliktCommand
 import com.github.ajalt.clikt.parameters.options.flag
 import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.parameters.options.required
-import io.dotanuki.arw.core.domain.errors.ErrorAware
-import io.dotanuki.arw.core.infrastructure.android.AndroidArtifactAnalyser
-import io.dotanuki.arw.core.infrastructure.cli.ErrorReporter
-import io.dotanuki.arw.core.infrastructure.filesystem.ValidatedFile
-import io.dotanuki.arw.features.common.ArtifactBaseline
-import java.io.File
+import io.dotanuki.arw.core.android.AndroidArtifactAnalyser
+import io.dotanuki.arw.core.cli.ErrorReporter
+import io.dotanuki.arw.core.errors.ErrorAware
+import io.dotanuki.arw.core.filesystem.ValidatedFile
+import io.dotanuki.arw.core.toml.ValidatedBaseline
 
 context (CompareContext)
 class CompareCommand : CliktCommand(
@@ -30,8 +29,7 @@ class CompareCommand : CliktCommand(
     context (ErrorAware)
     private fun performComparison() {
         val current = AndroidArtifactAnalyser.analyse(ValidatedFile(target))
-        val baselineFile = File(ValidatedFile(baseline))
-        val reference = tomlSerializer.decodeFromString(ArtifactBaseline.serializer(), baselineFile.readText())
+        val reference = ValidatedBaseline(ValidatedFile(baseline))
         val comparison = ArtifactsComparator.compare(current, reference.asArtifact())
         ComparisonReporter.reportChanges(comparison)
     }
