+++ src/test/kotlin/io/dotanuki/arw/features/comparison/ArtifactsComparatorTests.kt
@@ -2,13 +2,13 @@ package io.dotanuki.arw.features.comparison
 
 import arrow.core.raise.recover
 import com.google.common.truth.Truth.assertThat
-import io.dotanuki.arw.core.domain.models.AnalysedArtifact
-import io.dotanuki.arw.core.domain.models.AndroidComponent
-import io.dotanuki.arw.core.domain.models.AndroidComponentType.ACTIVITY
-import io.dotanuki.arw.core.domain.models.AndroidComponentType.APPLICATION
-import io.dotanuki.arw.core.domain.models.AndroidComponentType.PROVIDER
-import io.dotanuki.arw.core.domain.models.AndroidComponentType.RECEIVER
-import io.dotanuki.arw.core.domain.models.AndroidComponentType.SERVICE
+import io.dotanuki.arw.core.android.AnalysedArtifact
+import io.dotanuki.arw.core.android.AndroidComponent
+import io.dotanuki.arw.core.android.AndroidComponentType.ACTIVITY
+import io.dotanuki.arw.core.android.AndroidComponentType.APPLICATION
+import io.dotanuki.arw.core.android.AndroidComponentType.PROVIDER
+import io.dotanuki.arw.core.android.AndroidComponentType.RECEIVER
+import io.dotanuki.arw.core.android.AndroidComponentType.SERVICE
 import io.dotanuki.arw.helpers.errorAwareTest
 import org.junit.Test
